+++ app/support/DbAdapter/index.d.ts
@@ -232,8 +232,15 @@ export class DbAdapter {
       activityOnPropagable: boolean;
       activityHideIds: number[];
       authorsIds: ListLike<UUID>;
+      feedPinnedIn: UUID | null;
     }>,
   ): Promise<UUID[]>;
+  // Pinned posts
+  pinUserPost(feedId: UUID, postId: UUID, pinnedBy: UUID): Promise<boolean>;
+  unpinUserPost(feedId: UUID, postId: UUID): Promise<boolean>;
+  getPinnedDetailsByPosts(
+    postIds: UUID[],
+  ): Promise<Map<UUID, { userId: UUID; createdAt: string; pinnedBy: UUID }[]>>;
 
   // Likes
   unlikePost(postId: UUID, userId: UUID): Promise<boolean>;