+++ migrations/20251005182341_undo_delete_post_comments.ts
@@ -1,3 +1,9 @@
+// MIGRATION NAME ISSUE
+//
+// This migration was named incorrectly (it should be `20250510…` instead of
+// `20251005…`), but it has been applied to the database. So we need to keep the
+// following migrations names greater than it until October 5th 2025.
+
 import type { Knex } from 'knex';
 
 import { eventTypesSQLs } from '../app/support/migrations';